+++ src/Application/Category/Delete/CategoryDeleteService.php
@@ -6,16 +6,23 @@ namespace App\Application\Category\Delete;
 
 use App\Application\DTO;
 use App\Application\TransactionalService;
+use App\Domain\Activity\ActivityRepository;
+use App\Domain\Category\CategoryLinkedWithActivities;
 use App\Domain\Category\CategoryNotFound;
 use App\Domain\Category\CategoryRepository;
 
 class CategoryDeleteService implements TransactionalService
 {
     private CategoryRepository $repository;
+    /**
+     * @var ActivityRepository
+     */
+    private ActivityRepository $activityRepository;
 
-    public function __construct(CategoryRepository $repository)
+    public function __construct(CategoryRepository $repository, ActivityRepository $activityRepository)
     {
         $this->repository = $repository;
+        $this->activityRepository = $activityRepository;
     }
 
     public function __invoke(DTO $dto): DTO
@@ -24,6 +31,11 @@ class CategoryDeleteService implements TransactionalService
         $category = $this->repository->findOneById($dto->id());
         if(!$category){
             throw new CategoryNotFound($dto->id());
+        }
+
+        $activities = $this->activityRepository->findByCategory($category);
+        if(!empty($activities)){
+            throw new CategoryLinkedWithActivities($category->id());
         }
 
         $this->repository->remove($category);
